The module server for exploring prediction covariate summary results

The module server for exploring prediction covariate summary results

patientLevelPredictionCovariateSummaryServer(
  id,
  modelDesignId,
  developmentDatabaseId,
  performanceId,
  connectionHandler,
  inputSingleView,
  resultDatabaseSettings
)

Arguments

id

the unique reference id for the module

modelDesignId

unique id for the model design

developmentDatabaseId

unique id for the development database

performanceId

unique id for the performance results

connectionHandler

the connection to the prediction result database

inputSingleView

the current tab

resultDatabaseSettings

a list containing the result schema and prefixes

Value

The server to the covariate summary module

Details

The user specifies the id for the module

See also